pyproject.toml
[project]
name = "cv-compositor"
version = "0.1.0"
description = ""
authors = [
    {name = "bileamScheuvens",email = "benedictbileam@gmx.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "streamlit (>=1.49.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
